fix: scan AVG() result into float64, add regression test
SQLite AVG() returns float64 but AvgBytes was typed int64, causing a
scan error when the average was non-integer. Add a db-level test that
exercises this with an odd-sum pair to prevent regression.

+++ internal/db/queries.go
@@ -34,7 +34,7 @@ type AggregateRow struct {
 	ToolName   string
 	Calls      int64
 	TotalBytes int64
-	AvgBytes   int64
+	AvgBytes   float64
 }
 
 func InsertToolCall(db *sql.DB, tc ToolCall) error {
